Small refactor tests
